Fixed the window dragging awkwardly.
Sometimes, with 500 songs or so loaded and near the bottom, the player begins to not drag correctly due to lag I presume. For instance, say I hold M2 in the middle of the shuffle button, and when I drag it quickly across the screen, it goes way off. Off the entire window even, sometimes. In the end, it simply does not get dragged where I want it to go and this has become enough of a problem for me that I made this pull request. 

My first pull request, whoo.